# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SubSpace Subscription Marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  icon: string;
  description: string;
  serviceCount: number;
}

export interface SubscriptionService {
  id: number;
  name: string;
  description: string;
  logo: string;
  categoryId: number;
  categoryName: string;
  categorySlug: string;
  priceMonthly: number;
  priceAnnual?: number | null;
  currency: string;
  rating: number;
  reviewCount: number;
  subscriberCount: number;
  features: string[];
  tags: string[];
  isFeatured: boolean;
  website?: string | null;
  createdAt: string;
}

export interface CreateSubscriptionService {
  name: string;
  description: string;
  logo: string;
  categoryId: number;
  priceMonthly: number;
  priceAnnual?: number | null;
  currency: string;
  features: string[];
  tags: string[];
  website?: string | null;
}

export type UserSubscriptionPlanType =
  (typeof UserSubscriptionPlanType)[keyof typeof UserSubscriptionPlanType];

export const UserSubscriptionPlanType = {
  monthly: "monthly",
  annual: "annual",
} as const;

export type UserSubscriptionStatus =
  (typeof UserSubscriptionStatus)[keyof typeof UserSubscriptionStatus];

export const UserSubscriptionStatus = {
  active: "active",
  cancelled: "cancelled",
  paused: "paused",
} as const;

export interface UserSubscription {
  id: number;
  serviceId: number;
  serviceName: string;
  serviceLogo: string;
  planType: UserSubscriptionPlanType;
  price: number;
  currency: string;
  status: UserSubscriptionStatus;
  startedAt: string;
  renewsAt?: string | null;
  cancelledAt?: string | null;
  categoryName: string;
}

export type CreateUserSubscriptionPlanType =
  (typeof CreateUserSubscriptionPlanType)[keyof typeof CreateUserSubscriptionPlanType];

export const CreateUserSubscriptionPlanType = {
  monthly: "monthly",
  annual: "annual",
} as const;

export interface CreateUserSubscription {
  serviceId: number;
  planType: CreateUserSubscriptionPlanType;
}

export interface MarketplaceStats {
  totalServices: number;
  totalCategories: number;
  totalSubscribers: number;
  avgRating: number;
}

export type ListSubscriptionsParams = {
  category?: string;
  search?: string;
  minPrice?: number;
  maxPrice?: number;
  sortBy?: ListSubscriptionsSortBy;
};

export type ListSubscriptionsSortBy =
  (typeof ListSubscriptionsSortBy)[keyof typeof ListSubscriptionsSortBy];

export const ListSubscriptionsSortBy = {
  price: "price",
  name: "name",
  rating: "rating",
  newest: "newest",
} as const;