# lib/api-zod/src/generated/types/category.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SubSpace Subscription Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface Category {
  id: number;
  name: string;
  slug: string;
  icon: string;
  description: string;
  serviceCount: number;
}