# lib/api-zod/src/generated/types/createSubscriptionService.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SubSpace Subscription Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateSubscriptionService {
  name: string;
  description: string;
  logo: string;
  categoryId: number;
  priceMonthly: number;
  priceAnnual?: number | null;
  currency: string;
  features: string[];
  tags: string[];
  website?: string | null;
}