# lib/api-zod/src/generated/types/createUserSubscription.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SubSpace Subscription Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateUserSubscriptionPlanType } from "./createUserSubscriptionPlanType";

export interface CreateUserSubscription {
  serviceId: number;
  planType: CreateUserSubscriptionPlanType;
}