# lib/api-zod/src/generated/types/subscriptionService.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SubSpace Subscription Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface SubscriptionService {
  id: number;
  name: string;
  description: string;
  logo: string;
  categoryId: number;
  categoryName: string;
  categorySlug: string;
  priceMonthly: number;
  priceAnnual?: number | null;
  currency: string;
  rating: number;
  reviewCount: number;
  subscriberCount: number;
  features: string[];
  tags: string[];
  isFeatured: boolean;
  website?: string | null;
  createdAt: string;
}