# lib/api-zod/src/generated/types/userSubscription.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SubSpace Subscription Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { UserSubscriptionPlanType } from "./userSubscriptionPlanType";
import type { UserSubscriptionStatus } from "./userSubscriptionStatus";

export interface UserSubscription {
  id: number;
  serviceId: number;
  serviceName: string;
  serviceLogo: string;
  planType: UserSubscriptionPlanType;
  price: number;
  currency: string;
  status: UserSubscriptionStatus;
  startedAt: string;
  renewsAt?: string | null;
  cancelledAt?: string | null;
  categoryName: string;
}